checkRNG {rngtools}	R Documentation

Checking RNG Differences in Unit Tests

Description

checkRNG checks if two objects have the same RNG settings and should be used in unit tests, e.g., with the RUnit package.

Usage

checkRNG(x, y = getRNG(), ...)

Arguments

x, y
objects from which RNG settings are extracted.

...
extra arguments passed to checkTrue.

Examples

#--- checkRNG ---

# check for differences in RNG
set.seed(123)
checkRNG(123)
try( checkRNG(123, 123) )
try( checkRNG(123, 1:3) )

[Package rngtools version 1.4 Index]